update: change panel to frame
